pyproject.toml
[project]
name = "tickets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django~=5.2.3",
    "djangorestframework~=3.16.0",
    "drf-spectacular~=0.28.0",
    "drf-spectacular-sidecar~=2025.7.1",
    "psycopg[binary]~=3.2.9",
    "pyjwt~=2.10.1",
    "cryptography~=45.0.5",
    "opentelemetry-sdk~=1.34.1",
    "opentelemetry-api~=1.34.1",
    "opentelemetry-instrumentation-django~=0.55b1",
    "opentelemetry-instrumentation-requests~=0.55b1",
    "opentelemetry-instrumentation-logging~=0.55b1",
    "opentelemetry-exporter-zipkin~=1.11.1",
    "celery[redis]~=5.5.3",
    "django-celery-beat~=2.8.1",
    "flower~=2.0.1",
    "django-storages[boto3]~=1.14.6",
    "reportlab~=4.4.3",
    "django-unfold~=0.63.0",
    "django-environ~=0.12.0",
]

[dependency-groups]
test = [
    "coverage~=7.9.1",
    "pytest~=8.4.1",
    "pytest-cov~=6.2.1",
    "pytest-django~=4.11.1",
    "pytest-mock>=3.14.1",
]

lint = [
    "pre-commit~=4.2.0",
    "ruff~=0.12.3",
]

types = [
    "mypy~=1.17.0",
    "mypy-extensions~=1.1.0",
    "celery-stubs~=0.1.3",
    "boto3-stubs-lite~=1.39.4",
    "django-stubs~=5.2.1",
    "django-stubs-ext~=5.2.1",
    "djangorestframework-stubs~=3.16.1",
]

[tool.ruff]
line-length = 120
indent-width = 4
extend-exclude = ["migrations"]
lint.select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "ASYNC", # flake8-async
    "T10", # flake8-debugger
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "BLE", # flake8-blind-except
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
    "FIX", # flake8 - fix me
    "N", # pep8-naming
    "UP", # pyupgrade
]
lint.ignore = [
    "ISC001", # Disabled to avoid unexpected behaviour on `ruff format .`
    "B904",
    "BLE001"
]
format.quote-style = "double"
format.indent-style = "space"
format.skip-magic-trailing-comma = false
format.line-ending = "auto"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tickets.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]

[tool.coverage]
run.source = ["tickets"]
run.branch = true
run.omit = [
    "*/migrations/*",
    "*/__main__.py",
    "*/wsgi.py",
    "*/asgi.py",
    "*/celery.py",
]
report.show_missing = true
report.skip_covered = false
report.skip_empty = false
report.fail_under = 80

[tool.mypy]
mypy_path = [""]
files = ["tickets"]
python_version = "3.13"
explicit_package_bases = true
ignore_missing_imports = true
strict = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_unused_ignores = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main"
]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[[tool.mypy.overrides]]
module = "boto3.*"
ignore_missing_imports = true

[tool.django-stubs]
PYTHONPATH = ""
django_settings_module = "tickets.settings"
strict_settings = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
